requirements.txt
pandas
numpy
yfinance
sqlalchemy
psycopg2-binary
python-dotenv
jupyter
notebook
streamlit
matplotlib
plotly
scikit-learn
xgboost
apache-airflow
joblib
